backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
celery
redis
pypdf